// packages/bespoke/game/StockTrading/Home/src/interface.d.ts
export interface ICreateParams {

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);

    /** CreateParams playUrls */
    playUrls?: (string[]|null);
}
// ...
export interface IMoveParams {

    /** MoveParams stage */
    stage?: (number|null);
}
// ...
export interface IPushParams {
}
// ...
export interface IGameState {
}
// ...
export interface IPlayerState {

    /** PlayerState currentStage */
    currentStage?: (number|null);

    /** PlayerState unlockedStage */
    unlockedStage?: (number|null);
}